+++ components/core/tests/test-GrepCore.cpp
@@ -264,12 +264,12 @@ TEST_CASE("generate_logtype_string_for_empty_interpretation", "[dfa_search]") {
     auto const wildcard_encodable_positions{
             clp::GrepCoreTest::get_wildcard_encodable_positions(interpretation)
     };
-    size_t const num_combos{static_cast<size_t>(1) << wildcard_encodable_positions.size()};
+    uint64_t const num_combos{1ULL << wildcard_encodable_positions.size()};
     REQUIRE(1 == num_combos);
-    for (size_t mask{0}; mask < num_combos; ++mask) {
+    for (uint64_t mask{0}; mask < num_combos; ++mask) {
         std::unordered_map<size_t, bool> wildcard_mask_map;
         for (size_t i{0}; i < wildcard_encodable_positions.size(); ++i) {
-            wildcard_mask_map[wildcard_encodable_positions[i]] = mask >> i & 1;
+            wildcard_mask_map[wildcard_encodable_positions[i]] = mask >> i & 1ULL;
         }
         auto logtype_string{
                 clp::GrepCoreTest::generate_logtype_string(interpretation, wildcard_mask_map)
